\name{MultiwayRegression-package}
\alias{MultiwayRegression}
\docType{package}
\title{
Perform tensor-on-tensor regression
}
\description{
Functions to predict one multi-way array (i.e., a tensor) from another multi-way array. 
}
\details{
\tabular{ll}{
Package: \tab MultiwayRegression-package\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-12-31\cr
License: \tab GPL-3\cr
}

}
\author{
Eric F. Lock

Maintainer: Eric F. Lock <elock@umn.edu>
}

\examples{

}


\keyword{ package }