\name{SimData}
\alias{SimData}
\docType{data}
\title{Simulated multi-way data for prediction}
\description{
Simulated multi-way data for prediction.        
}
\usage{data(IFNB_Data)}
\format{
\itemize{
\item X: predictor array of dimension 100 x 15 x 20  
\item Y: outcome array of dimension 100 x 5 x 10  
}}
   

\keyword{datasets}
